Changed route script into a board controller
Deleted the gizmos of the board
