--- ChutesAndLaddersUniRX/Assets/_MAIN/Scripts/Board/Route.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12653d2d15f064142a8e1aa27a57a9cf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
